pyproject.toml
[project]
name = "complainceqapipline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "azure-identity>=1.25.3",
    "azure-monitor-opentelemetry>=1.8.6",
    "azure-search-documents>=11.6.0",
    "azure-storage-blob>=12.28.0",
    "fastapi>=0.135.1",
    "firecrawl-py>=4.19.0",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "langsmith>=0.7.17",
    "opentelemetry-instrumentation-fastapi>=0.60b0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "redis>=7.3.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
    "uvicorn>=0.41.0",
    "yt-dlp>=2026.3.13",
]
